Added date to tariff table and calculation

+++ src/main/java/com/lynus/cs203/dtos/response/TariffRateResponse.java
@@ -47,5 +47,5 @@ public class TariffRateResponse {
     @Schema(
             description = "Tariff rate value"
     )
-    private int value;
+    private double value;
 }
